+++ common/src/main/java/raccoonman/reterraforged/world/worldgen/RTFRandomState.java
@@ -15,9 +15,6 @@ public interface RTFRandomState {
 
 	@Nullable
 	GeneratorContext generatorContext();
-
-	@Nullable
-	DensityFunction openClFinalDensity();
 	
 	DensityFunction wrap(DensityFunction function);